test: Add assertion for inline comment updating
